refactor: renamed types file
